requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
attrs==24.3.0
blinker==1.8.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distro==1.9.0
et-xmlfile==2.0.0
exceptiongroup==1.2.2
flask==3.0.3
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.5
jiter==0.8.2
markupsafe==2.1.5
multidict==6.1.0
nest-asyncio==1.6.0
numpy==1.24.4
openai==0.28.0
openpyxl==3.1.5
packaging==24.2
pandas==2.0.3
plotly==5.24.1
pluggy==1.5.0
propcache==0.2.0
pydantic==2.10.4
pydantic-core==2.27.2
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
retrying==1.3.4
setuptools==75.3.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.36
tenacity==9.0.0
tomli==2.2.1
tqdm==4.67.1
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
werkzeug==3.0.6
yarl==1.15.2
zipp==3.20.2
